pyproject.toml
[project]
name = "manage-item-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache[redis]>=0.12.3",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.1",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.46",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "python-json-logger>=4.0.0",
]
